+++ eNMS/controller/base.py
@@ -425,7 +425,6 @@ class BaseController:
             if not value:
                 continue
             filter = kwargs["form"].get(f"{property}_filter")
-            print(kwargs["form"])
             if value in ("bool-true", "bool-false"):
                 constraint = getattr(model, property) == (value == "bool-true")
             elif filter == "equality":