+++ tests/test_inventory.py
@@ -2,6 +2,7 @@ from werkzeug.datastructures import ImmutableMultiDict
 
 from eNMS import app
 from eNMS.database.functions import delete_all, fetch, fetch_all
+from eNMS.settings import pool_properties
 
 from tests.conftest import check_pages
 
@@ -14,6 +15,7 @@ def define_device(number: str, description: str) -> ImmutableMultiDict:
             ("description", description),
             ("location", "paris"),
             ("vendor", "Cisco"),
+            ("icon", "router"),
             ("operating_system", "IOS"),
             ("os_version", "1.4.4.2"),
             ("longitude", "12"),
@@ -27,7 +29,7 @@ def define_link(source: int, destination: int) -> ImmutableMultiDict:
     return ImmutableMultiDict(
         [
             ("form_type", "link"),
-            ("name", f"{source} to {destination}"),
+            ("name", f"{source} too {destination}"),
             ("description", "description"),
             ("location", "Los Angeles"),
             ("vendor", "Juniper"),
@@ -49,7 +51,7 @@ def test_manual_object_creation(user_client):
         for destination in devices[:3]:
             obj_dict = define_link(source.id, destination.id)
             user_client.post("/update/link", data=obj_dict)
-    assert len(fetch_all("device")) == 20
+    assert len(fetch_all("device")) == 10
     assert len(fetch_all("link")) == 9
 
 
@@ -117,11 +119,21 @@ pool2 = {
 }
 
 
+def create_pool(pool: dict) -> dict:
+    for property in pool_properties["device"]:
+        if f"device_{property}_match" not in pool:
+            pool[f"device_{property}_match"] = "inclusion"
+    for property in pool_properties["link"]:
+        if f"link_{property}_match" not in pool:
+            pool[f"link_{property}_match"] = "inclusion"
+    return pool
+
+
 @check_pages("table/device", "table/link", "view/network")
 def test_pool_management(user_client):
     create_from_file(user_client, "europe.xls")
-    user_client.post("/update/pool", data=pool1)
-    user_client.post("/update/pool", data=pool2)
+    user_client.post("/update/pool", data=create_pool(pool1))
+    user_client.post("/update/pool", data=create_pool(pool2))
     p1, p2 = fetch("pool", name="pool1"), fetch("pool", name="pool2")
     assert len(p1.devices) == 21
     assert len(p1.links) == 20