+++ docs/base/release_notes.rst
@@ -35,6 +35,7 @@ Version 3.19
 - New Table properties mechanism: all table properties are displayed in a JSON file: you can configure which ones
 appear in each table by default, whether they are searchable or not, etc, their label in the UI, etc.
 You will need to add your CUSTOM properties to that file if you want them to appear in the table.
+- Same with dashboard properties and pool properties
 - New Column visibility feature
 - New Configuration Management Mechanism